set default size of Aint to integer*8 ... not ideal, though. Real mpich testsuite uses autoconf to configure the size

+++ teshsuite/smpi/mpich3-test/f77/rma/addsize.h
@@ -3,4 +3,4 @@ C
 C  (C) 2003 by Argonne National Laboratory.
 C      See COPYRIGHT in top-level directory.
 C
-      integer asize
+      integer*8 asize
